backend/pyproject.toml
[project]
name = "step1x3d-backend"
version = "2.0.0"
description = "FastAPI backend for Step1X-3D local GPU inference"
authors = [
    {name = "HPE Team", email = "team@hpe.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pillow>=10.1.0",
    "requests>=2.31.0",
    "trimesh>=4.0.0",
    "nibabel>=5.1.0",
    "numpy>=1.24.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.36.0",
    "diffusers>=0.24.0",
    "accelerate>=0.25.0",
    "xformers>=0.0.22",
    "scikit-image>=0.22.0",
    "aiofiles>=23.2.1",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
